+++ pastel/dupe_detection_test.go
@@ -59,7 +59,7 @@ func TestToCompressSignedDDAndFingerprints(t *testing.T) {
 					Sexy:     0.5,
 				},
 
-				ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+				ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 				HashOfCandidateImageFile: "HashOfCandidateImageFile",
 			},
@@ -144,7 +144,7 @@ func TestGetIDFiles(t *testing.T) {
 					Sexy:     0.5,
 				},
 
-				ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+				ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 				HashOfCandidateImageFile: "HashOfCandidateImageFile",
 			},
