+++ walletnode/services/senseregister/task_test.go
@@ -199,7 +199,7 @@ func TestTaskRun(t *testing.T) {
 					Sexy:     0.5,
 				},
 
-				ImageFingerprintOfCandidateImageFile: []float32{1, 2, 3},
+				ImageFingerprintOfCandidateImageFile: []float64{1, 2, 3},
 
 				HashOfCandidateImageFile: "HashOfCandidateImageFile",
 			}
